Fix labels being floats, improve testing

+++ asreview/review/oracle.py
@@ -108,7 +108,7 @@ class ReviewOracle(BaseReview):
 
     def _get_labels(self, ind):
 
-        y = np.zeros((len(ind), ))
+        y = np.zeros((len(ind), ), dtype=int)
 
         for j, index in enumerate(ind):
